+++ src/dice/damage-dialog.ts
@@ -951,7 +951,7 @@ function createDamageCardContent(
           </div>
           ${selectedPower.specials && selectedPower.specials.length > 0 ? `
             <div class="damage-row">
-              <span class="damage-label">Power Specials:</span>
+              <span class="damage-label">Power Special Effects:</span>
               <span class="damage-value">${selectedPower.specials.join(', ')}</span>
             </div>
           ` : ''}